requirements.txt
boto3
gradio
griptape
python-dotenv
trafilatura
